requirements.txt
# requirements.txt - Python Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.7
gunicorn==21.2.0

# For production deployment
supervisor==4.2.5